refactor: Move email account related code to email_account.py
